+++ src/integration-test/java/org/openlmis/notification/web/BaseWebIntegrationTest.java
@@ -29,21 +29,20 @@ import com.github.tomakehurst.wiremock.junit.WireMockRule;
 import com.jayway.restassured.RestAssured;
 import com.jayway.restassured.config.ObjectMapperConfig;
 import com.jayway.restassured.config.RestAssuredConfig;
-
+import com.jayway.restassured.specification.RequestSpecification;
+import guru.nidi.ramltester.RamlDefinition;
+import guru.nidi.ramltester.RamlLoaders;
+import guru.nidi.ramltester.restassured.RestAssuredClient;
+import javax.annotation.PostConstruct;
 import org.junit.Rule;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.embedded.LocalServerPort;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.HttpHeaders;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
-
-import guru.nidi.ramltester.RamlDefinition;
-import guru.nidi.ramltester.RamlLoaders;
-import guru.nidi.ramltester.restassured.RestAssuredClient;
-
-import javax.annotation.PostConstruct;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
@@ -115,6 +114,12 @@ public abstract class BaseWebIntegrationTest {
    */
   protected String getTokenHeader() {
     return USER_ACCESS_TOKEN_HEADER;
+  }
+
+  protected RequestSpecification startRequest() {
+    return restAssured
+        .given()
+        .header(HttpHeaders.AUTHORIZATION, USER_ACCESS_TOKEN_HEADER);
   }
 
 }