+++ src/main/java/org/openlmis/notification/web/usercontactdetails/EmailVerificationTokenDto.java
@@ -13,25 +13,23 @@
  * http://www.gnu.org/licenses.  For additional information contact info@OpenLMIS.org. 
  */
 
-package org.openlmis.notification.service.referencedata;
+package org.openlmis.notification.web.usercontactdetails;
 
+import java.time.ZonedDateTime;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
-import org.openlmis.notification.web.BaseDto;
 
 @Getter
 @Setter
-@ToString
 @NoArgsConstructor
 @AllArgsConstructor
-@EqualsAndHashCode(callSuper = true)
-public final class UserDto extends BaseDto {
-  private String username;
-  private String firstName;
-  private String lastName;
-  private boolean active;
+@EqualsAndHashCode
+@ToString
+public final class EmailVerificationTokenDto {
+  private String emailAddress;
+  private ZonedDateTime expiryDate;
 }